Upgrade arg parser to use subparsers
The previous implementation of the argument parsing in `bin/plunchy`
took a shortcut that is non-optimal. The old way forced me to do some
validation in the specific command handlers, when that validation could
and should have been done by argparse.

Other changes to the code were made to accomodate the subparsers.
Previously, everything used `self.arg`, but is now split into
`self.pattern` and `self.path`, with each being used correctly by the
sub-commands.

I also removed the `add` command in favor of only having an `install`
command.