Remove CLI script in favor of setup.py entry point
There's no real need for the CLI script to be there, when a simple entry
point will do just fine. The CLI script didn't really do anything except
setup the parser and pass the args to Plunchy, and that can happen in a
function in the plunchy module just as easily.